% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{a3}
\alias{a3}
\title{Create an \code{a3} object}
\usage{
a3(
  seq,
  site = NULL,
  region = NULL,
  ptm = NULL,
  clv = NULL,
  variant = NULL,
  uniprotid = NULL,
  description = NULL,
  reference = NULL
)
}
\arguments{
\item{seq}{Character: Amino acid sequence.}

\item{site}{Named list of vectors of integer indices of sites, e.g.
\code{list("N-terminal repeat" = c(46, 47, 52), "Microtubule binding domain" = c(244, 245, 246))}}

\item{region}{Named list of integer indices,
e.g. \code{list("Phosphodegron" = c(46, 47, 48, 49, 50, 51), "KXGS" = c(259, 260, 261, 262))}
or character vectors with index range of regions in format
\code{start:end}, e.g. \code{list(Phosphodegron = c("46:51", "149:154"), KXGS = c("259:262", "290:293"))}}

\item{ptm}{Named list of vectors with indices of post-translational modifications, e.g.
\code{list("Phosphorylation" = c(17, 18, 29, 30), "Acetylation" = c(148, 150, 163))}}

\item{clv}{Named list of cleavage sites, e.g.
\code{list(CTSL = c(54, 244, 319), CTSD = c(340, 391, 426))}}

\item{variant}{List of lists with variant information. Each list must contain a
\code{Position} element}

\item{uniprotid}{Character: Uniprot ID.}

\item{description}{Character: Description of the data / experiment.}

\item{reference}{Character: Link to reference (journal publication, preprint, etc.)}
}
\value{
\code{a3} object
}
\description{
Creates an \code{a3} object given amino acid sequence and annotations.
}
\details{
We choose to keep NULL elements as empty lists in JSON, since we want users to be
able to easily add annotations, whether programmaticaly, using a web app, or
manually.
}
\author{
EDG
}
