% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_sub.R
\name{aa_sub}
\alias{aa_sub}
\title{Perform amino acid substitutions}
\usage{
aa_sub(x, substitutions, verbosity = 1)
}
\arguments{
\item{x}{Character or charactr vector: Amino acid sequence. e.g. \code{"ARND"} or
\code{c("A", "R", "N", "D")}.}

\item{substitutions}{Character vector: Substitutions to perform in the format
"OriginalPositionNew", e.g. \code{c("C291A", "C322A")}.}
}
\value{
Character vector with substitutions performed.
}
\description{
Perform amino acid substitutions
}
\author{
EDG
}
