% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt.R
\name{aggregate.xt}
\alias{aggregate.xt}
\title{Aggregate method for \code{xt} object}
\usage{
\method{aggregate}{xt}(x, group, fn = mean)
}
\arguments{
\item{x}{\code{xt} object.}

\item{group}{Character: Grouping variable.}

\item{fn}{Function: Function to apply to each group.}
}
\description{
Aggregate method for \code{xt} object
}
\author{
EDG
}
