% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{as.a3.list}
\alias{as.a3.list}
\title{as.a3.list method}
\usage{
\method{as.a3}{list}(x)
}
\arguments{
\item{x}{List: Named list with elements \code{Sequence}, \code{Annotations}, \code{UniprotID}.
\code{Annotations} is a named list with possible elements \code{Site}, \code{Region}, \code{PTM},
\code{Cleavage_site}, \code{Variant}, \code{Description}, \code{Reference}.}
}
\value{
\code{a3} object.
}
\description{
as.a3.list method
}
\author{
EDG
}
