% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt.R
\name{as.xt.list}
\alias{as.xt.list}
\title{as.xt.list method}
\usage{
\method{as.xt}{list}(x)
}
\arguments{
\item{x}{List: Named list with elements \code{x}, \code{y}, \code{y2}, \code{xunits}, \code{yunits}, \code{y2units},
\code{Description}, \code{Reference}.}
}
\value{
\code{xt} object.
}
\description{
as.xt.list method
}
\author{
EDG
}
