% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_A3.R
\name{as_A3}
\alias{as_A3}
\title{as_A3}
\usage{
as_A3(x, ...)
}
\arguments{
\item{x}{List: Named list with elements \code{Sequence}, \code{Annotations}, \code{UniprotID}.
\code{Annotations} is a named list with possible elements \code{Site}, \code{Region}, \code{PTM},
\code{Cleavage_site}, \code{Variant}, \code{Description}, \code{Reference}.}
}
\value{
\code{A3} object.
}
\description{
as_A3
}
\author{
EDG
}
