% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_metrics.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Cosine similarity}
\usage{
cosine_similarity(x, y, sequence)
}
\arguments{
\item{x}{Character vector. Will be coerced to character.}

\item{y}{Character vector. Will be coerced to character.}
}
\value{
Numeric: Cosine similarity.
}
\description{
Cosine similarity
}
\author{
EDG
}
