% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_metrics.R
\name{dice_coefficient}
\alias{dice_coefficient}
\title{Dice Coefficient}
\usage{
dice_coefficient(x, y)
}
\arguments{
\item{x}{Character vector. Will be coerced to character.}

\item{y}{Character vector. Will be coerced to character.}
}
\value{
Numeric: Dice coefficient.
}
\description{
Dice Coefficient
}
\author{
EDG
}
