% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2sequence.R
\name{gene2sequence}
\alias{gene2sequence}
\title{Get the sequence of a gene}
\usage{
gene2sequence(
  gene,
  organism = "hsapiens",
  biomart = "ensembl",
  host = "https://www.ensembl.org",
  seq_type = "coding",
  verbosity = 1
)
}
\arguments{
\item{gene}{Character: Gene name.}

\item{organism}{Character: Organism name.}

\item{biomart}{Character: Biomart name.}

\item{host}{Character: Host address.}
}
\value{
data.frame with columns "gene", "ensembl_transcript_id" and "sequence".
}
\description{
Get the sequence of a gene
}
\author{
EDG
}
