% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_utils.R
\name{get_alphafold}
\alias{get_alphafold}
\title{Get AlphaFold info for a given UniProt ID}
\usage{
get_alphafold(uniprotid)
}
\arguments{
\item{uniprotid}{Character: UniProt ID.}
}
\value{
data frame with AlphaFold info.
}
\description{
Get AlphaFold info for a given UniProt ID
}
\author{
EDG
}
