% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{int2range}
\alias{int2range}
\title{Convert integer range to character with colon separator}
\usage{
int2range(x)
}
\arguments{
\item{x}{Integer vector. Must be consecutive integers from lowest to highest.}
}
\value{
Character with colon separator.
}
\description{
Convert integer range to character with colon separator
}
\examples{
\dontrun{
x <- 34:42
int2range(x)
int2range(28:34)
int2range(c(3, 4, 5, 6))
# This will throw an error:
int2range(c(3, 4, 5, 6, 8))
}
}
\author{
EDG
}
