% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_metrics.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Jaccard index}
\usage{
jaccard_index(x, y)
}
\arguments{
\item{x}{Character vector. Will be coerced to character.}

\item{y}{Character vector. Will be coerced to character.}
}
\value{
Numeric: Jaccard index.
}
\description{
Jaccard index
}
\author{
EDG
}
