% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_metrics.R
\name{overlap_coefficient}
\alias{overlap_coefficient}
\title{Overlap Coefficient}
\usage{
overlap_coefficient(x, y)
}
\arguments{
\item{x}{Character vector. Will be coerced to character.}

\item{y}{Character vector. Will be coerced to character.}
}
\value{
Numeric: Overlap coefficient.
}
\description{
Overlap Coefficient a.k.a. Szymkiewicz-Simpson coefficient.
}
\author{
EDG
}
