% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_metrics.R
\name{pairwise_similarity}
\alias{pairwise_similarity}
\title{Pairwise PTM similarity}
\usage{
pairwise_similarity(x, metric = "jaccard", sequence = NULL)
}
\arguments{
\item{x}{List of PTM annotations}

\item{metric}{Character: "jaccard", "dice", "overlap", "cosine", "pmi"}
}
\value{
Matrix of pairwise similarity scores.
}
\description{
Pairwise PTM similarity
}
\author{
EDG
}
