% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{plot.a3}
\alias{plot.a3}
\title{Plot method for \code{a3} object}
\usage{
\method{plot}{a3}(x, ...)
}
\arguments{
\item{x}{\code{a3} object.}

\item{...}{Additional arguments passed to \code{rtemis::dplot3_protein}.}
}
\description{
Plot method for \code{a3} object
}
\author{
EDG
}
