% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_metrics.R
\name{pmi}
\alias{pmi}
\title{Pointwise Mutual Information}
\usage{
pmi(x, y, data)
}
\arguments{
\item{x}{Character vector. Will be coerced to character.}

\item{y}{Character vector. Will be coerced to character.}
}
\value{
Numeric: Pointwise Mutual Information.
}
\description{
Pointwise Mutual Information
}
\author{
EDG
}
