% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{print.a3}
\alias{print.a3}
\title{Print method for \code{a3} object}
\usage{
\method{print}{a3}(x, head.n = 10, ...)
}
\arguments{
\item{x}{\code{a3} object.}

\item{...}{Not used.}
}
\description{
Print method for \code{a3} object
}
\author{
EDG
}
