% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt.R
\name{print.xt}
\alias{print.xt}
\title{Print method for \code{xt} object}
\usage{
\method{print}{xt}(x, head.n = 10, ...)
}
\arguments{
\item{x}{\code{xt} object.}

\item{...}{Not used.}
}
\value{
\code{xt} object, invisibly.
}
\description{
Print method for \code{xt} object
}
\author{
EDG
}
