% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.a3json.R
\name{read.a3json}
\alias{read.a3json}
\title{Read \code{a3} object from JSON file}
\usage{
read.a3json(filepath)
}
\arguments{
\item{filepath}{Character: Path to JSON file.}
}
\value{
\code{a3} object.
}
\description{
Read \code{a3} object from JSON file
}
\author{
EDG
}
