% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtjson.R
\name{read.xtjson}
\alias{read.xtjson}
\title{Read \code{xt} object from JSON file}
\usage{
read.xtjson(filepath, verbosity = 0L)
}
\arguments{
\item{filepath}{Character: Path to JSON file.}

\item{verbosity}{Integer: if greater than 0, print messages.}
}
\value{
\code{xt} object.
}
\description{
Read \code{xt} object from JSON file
}
\details{
Note that factors saved under \code{group} are written as character by \link{write.xtjson} and
when they are read back in, they are converted back to factors using \link{factor}. This means that
the levels will be set alphabetically. If needed, reorder them after reading in the JSON file
using \link{factor}.
}
\author{
EDG
}
