% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A3_utils.R
\name{read_A3json}
\alias{read_A3json}
\title{Read \code{A3} object from JSON file}
\usage{
read_A3json(filepath, verbosity = 0L)
}
\arguments{
\item{filepath}{Character: Path to JSON file.}

\item{verbosity}{Integer: if greater than 0, print messages.}
}
\value{
\code{A3} object.
}
\description{
Read \code{A3} object from JSON file
}
\author{
EDG
}
