% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rtemisbio-package}
\alias{rtemisbio}
\alias{rtemisbio-package}
\title{\pkg{rtemisbio}: Bioinformatics ops}
\description{
Bioinformatics utilities
}
\seealso{
Useful links:
\itemize{
  \item \url{https://egenn.github.io/rtemisbio/}
}

}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID})

}
