% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{summary.a3}
\alias{summary.a3}
\title{Summary method for \code{a3} object}
\usage{
\method{summary}{a3}(object, ...)
}
\arguments{
\item{object}{\code{a3} object.}
}
\description{
Summary method for \code{a3} object
}
\author{
EDG
}
