% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt.R
\name{toxt}
\alias{toxt}
\title{Create an \code{xt} object}
\usage{
toxt(
  x,
  y,
  x2 = NULL,
  y2 = NULL,
  group = NULL,
  shade = NULL,
  xunits = NULL,
  yunits = NULL,
  y2units = NULL,
  reference = NULL
)
}
\arguments{
\item{x}{Named list of datetime vectors.}

\item{y}{Named list of numeric vectors: When plotted, these will correspond to the
left y-axis.}

\item{y2}{Named list of numeric vectors: When plotted, these will correspond to the
right y-axis.}

\item{group}{Named list of factors: Grouping variable(s).}

\item{xunits}{Character: Units for \code{x}.}

\item{yunits}{Character: Units for \code{y}.}

\item{y2units}{Character: Units for \code{y2}.}

\item{reference}{Character: Link to reference (journal publication, preprint, etc.)}
}
\value{
\code{xt} object
}
\description{
Creates an \code{xt} object from time series data.
}
\author{
EDG
}
