% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.a3json.R
\name{write.a3json}
\alias{write.a3json}
\title{Write \code{a3} object to JSON file}
\usage{
write.a3json(x, filepath, overwrite = FALSE)
}
\arguments{
\item{x}{\code{a3} object, as created by \code{as.a3()}.}

\item{filepath}{Character: Path to save JSON file.}
}
\description{
Write \code{a3} object to JSON file
}
\author{
EDG
}
