% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtjson.R
\name{write.xtjson}
\alias{write.xtjson}
\title{Write \code{xt} object to JSON file}
\usage{
write.xtjson(x, filepath, overwrite = FALSE)
}
\arguments{
\item{x}{\code{xt} object, as created by \link{toxt} or \link{as.xt}.}

\item{filepath}{Character: Path to save JSON file.}
}
\description{
Write \code{xt} object to JSON file
}
\author{
EDG
}
