% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A3json.R
\name{write_A3json}
\alias{write_A3json}
\title{Write \code{A3} object to JSON file}
\usage{
write_A3json(x, filepath, overwrite = FALSE)
}
\arguments{
\item{x}{\code{A3} object, as created by \code{as_A3()}.}

\item{filepath}{Character: Path to save JSON file.}

\item{overwrite}{Logical: If TRUE, overwrite existing file.}
}
\value{
Nothing. Writes JSON file.
}
\description{
Write \code{A3} object to JSON file
}
\author{
EDG
}
