% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xt_utils.R
\name{write_Xtjson}
\alias{write_Xtjson}
\title{Write \code{Xt} object to JSON file}
\usage{
write_Xtjson(x, filepath, overwrite = FALSE)
}
\arguments{
\item{x}{\code{Xt} object, as created by \link{as_Xt}.}

\item{filepath}{Character: Path to save JSON file.}

\item{overwrite}{Logical: If TRUE, overwrite existing file.}
}
\value{
Nothing. Writes JSON file.
}
\description{
Write \code{Xt} object to JSON file
}
\author{
EDG
}
