% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xt.R
\name{Xt}
\alias{Xt}
\title{Xt Class}
\usage{
Xt(
  x = list(),
  y = list(),
  x2 = list(),
  y2 = list(),
  zt = integer(0),
  shade = integer(0),
  group = list(),
  xunits = character(0),
  yunits = character(0),
  y2units = character(0),
  description = character(0),
  reference = character(0)
)
}
\description{
Timeseries class is designed to store and manipulate time series data.
}
\section{Fields}{

\describe{
\item{\code{x}}{Named list of datetime vectors.}

\item{\code{y}}{Named list of numeric vectors: When plotted, these will correspond to the
left y-axis.}

\item{\code{x2}}{Named list of datetime vectors: When plotted, these will correspond to the
right x-axis. If not provided, \code{x} will be used for both \code{y} and \code{y2}.}

\item{\code{y2}}{Named list of numeric vectors: When plotted, these will correspond to the
right y-axis.}

\item{\code{zt}}{Numeric vector: Zeitgeber time. If provided, this will be used to label x-axis ticks.
Assumes a single datetime vector in \code{x}. Elements in \code{zt} must correspond to elements in \code{x}.}

\item{\code{shade}}{Binary vector: \code{0} indicates no shading, \code{1} indicates shading. If provided,
this will be used to shade the plot.}

\item{\code{group}}{Named list of factors: Grouping variable(s).}

\item{\code{xunits}}{Character: Units for \code{x}.}

\item{\code{yunits}}{Character: Units for \code{y}.}

\item{\code{y2units}}{Character: Units for \code{y2}.}

\item{\code{description}}{Character: Description of the data / experiment.}

\item{\code{reference}}{Character: Link to reference (journal publication, preprint, etc.)}
}}

