% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt.R
\name{aggregate.xt}
\alias{aggregate.xt}
\title{Aggregate method for \code{xt} object}
\usage{
\method{aggregate}{xt}(
  x,
  groupname,
  fn = mean,
  backend = getOption("rt.backend", "base"),
  ...
)
}
\arguments{
\item{x}{\code{xt} object.}

\item{fn}{Function: Function to apply to each group.}

\item{backend}{Character: "base", "data.table", or "dplyr"; backend to use for aggregation.}

\item{...}{Additional arguments passed to \code{fn}.}

\item{group}{Character: Grouping variable.}
}
\description{
Aggregate method for \code{xt} object
}
\author{
EDG
}
