% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Xt.R
\name{aggregate.Xt}
\alias{aggregate.Xt}
\title{Aggregate method for \code{Xt} object}
\usage{
\method{aggregate}{Xt}(
  x,
  groupname,
  fn = mean,
  backend = getOption("rtemis_backend", "base"),
  ...
)
}
\arguments{
\item{x}{\code{Xt} object.}

\item{groupname}{Character: Grouping variable.}

\item{fn}{Function: Function to apply to each group.}

\item{backend}{Character: "base", "data.table", or "dplyr"; backend to use for aggregation.}

\item{...}{Additional arguments passed to \code{fn}.}
}
\description{
Aggregate method for \code{Xt} object
}
\author{
EDG
}
