% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_A3.R
\name{plot.A3}
\alias{plot.A3}
\title{Plot method for \code{A3} object}
\usage{
\method{plot}{A3}(x, ...)
}
\arguments{
\item{x}{\code{A3} object.}

\item{...}{Additional arguments passed to \link[rtemis:draw_protein]{rtemis::draw_protein}.}
}
\value{
\code{plotly} object.
}
\description{
Plot method for \code{A3} object
}
\author{
EDG
}
