% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt.R
\name{plot.xt}
\alias{plot.xt}
\title{Plot method for \code{xt} object}
\usage{
\method{plot}{xt}(x, ...)
}
\arguments{
\item{x}{\code{xt} object.}

\item{...}{Additional arguments passed to \link[rtemis:dplot3_xt]{rtemis::dplot3_xt}.}
}
\description{
Plot method for \code{xt} object
}
\author{
EDG
}
