% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Xt.R
\name{plot.Xt}
\alias{plot.Xt}
\title{Plot method for \code{Xt} object}
\usage{
\method{plot}{Xt}(x, ...)
}
\arguments{
\item{x}{\code{Xt} object.}

\item{...}{Additional arguments passed to \link[rtemis:draw_xt]{rtemis::draw_xt}.}
}
\value{
\code{plotly} object.
}
\description{
Plot method for \code{Xt} object
}
\author{
EDG
}
