% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3_metrics.R
\name{pmi}
\alias{pmi}
\title{Pointwise Mutual Information}
\usage{
pmi(x, y, sequence)
}
\arguments{
\item{x}{Character vector of PTM sites. Will be coerced to character.}

\item{y}{Character vector of PTM sites. Will be coerced to character.}

\item{sequence}{Character vector representing the protein sequence, used to determine the total number of possible sites.}
}
\value{
Numeric: Pointwise Mutual Information.
}
\description{
Pointwise Mutual Information
}
\author{
EDG
}
