% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A3.R
\name{print.A3}
\alias{print.A3}
\title{Print method for \code{A3} object}
\usage{
\method{print}{A3}(x, head_n = 10, ...)
}
\arguments{
\item{x}{\code{A3} object.}

\item{head_n}{Integer: Number of characters to show from the sequence.}

\item{...}{Not used.}
}
\value{
Called for side effects, prints object to console
}
\description{
Print method for \code{A3} object
}
\author{
EDG
}
