% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtjson.R
\name{read.xtjson}
\alias{read.xtjson}
\title{Read \code{xt} object from JSON file}
\usage{
read.xtjson(filepath, verbosity = 0L)
}
\arguments{
\item{filepath}{Character: Path to JSON file.}

\item{verbosity}{Integer: if greater than 0, print messages.}
}
\value{
\code{xt} object.
}
\description{
Read \code{xt} object from JSON file
}
\author{
EDG
}
