% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis.bio-package.R
\docType{package}
\name{rtemis.bio-package}
\alias{rtemis.bio}
\alias{rtemis.bio-package}
\alias{rtemis.bio-package}
\title{rtemis.bio: Bioinformatics ops}
\description{
Bioinformatics utilities
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rtemis.bio.rtemis.org}
}

}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID})

}
