% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A3.R
\name{summary.A3}
\alias{summary.A3}
\title{Summary method for \code{A3} object}
\usage{
\method{summary}{A3}(object, ...)
}
\arguments{
\item{object}{\code{A3} object.}

\item{...}{Not used}
}
\value{
Called for side effects, prints summary to console.
}
\description{
Summary method for \code{A3} object
}
\author{
EDG
}
